refactor: discovery
