+++ apps/wallet-mobile/src/features/Discover/useCases/ReviewTransaction/ReviewTransaction.tsx
@@ -1,4 +1,3 @@
-import {Transaction} from '@emurgo/cross-csl-core'
 import {createTypeGuardFromSchema, isNonNullable, truncateString} from '@yoroi/common'
 import {useTheme} from '@yoroi/theme'
 import {Portfolio} from '@yoroi/types'
@@ -8,7 +7,7 @@ import {useEffect} from 'react'
 import {StyleSheet, View} from 'react-native'
 import {TouchableOpacity} from 'react-native-gesture-handler'
 import {SafeAreaView} from 'react-native-safe-area-context'
-import {useMutation, useQuery} from 'react-query'
+import {useQuery} from 'react-query'
 import {z} from 'zod'
 
 import {Button} from '../../../../components/Button/Button'
@@ -17,37 +16,22 @@ import {Icon} from '../../../../components/Icon'
 import {ScrollView} from '../../../../components/ScrollView/ScrollView'
 import {Spacer} from '../../../../components/Spacer/Spacer'
 import {Text} from '../../../../components/Text'
-import {logger} from '../../../../kernel/logger/logger'
 import {useParams} from '../../../../kernel/navigation'
-import {cip30LedgerExtensionMaker} from '../../../../yoroi-wallets/cardano/cip30/cip30-ledger'
 import {wrappedCsl} from '../../../../yoroi-wallets/cardano/wrappedCsl'
 import {formatAdaWithText, formatTokenWithText} from '../../../../yoroi-wallets/utils/format'
 import {asQuantity} from '../../../../yoroi-wallets/utils/utils'
 import {usePortfolioTokenInfos} from '../../../Portfolio/common/hooks/usePortfolioTokenInfos'
 import {useSelectedWallet} from '../../../WalletManager/common/hooks/useSelectedWallet'
-import {useConfirmHWConnectionModal} from '../../common/ConfirmHWConnectionModal'
-import {isUserRejectedError, userRejectedError} from '../../common/errors'
-import {usePromptRootKey} from '../../common/hooks'
 import {useStrings} from '../../common/useStrings'
 
-export type ReviewTransactionParams =
-  | {
-      isHW: false
-      cbor: string
-      onConfirm: (rootKey: string) => void
-      onCancel: () => void
-    }
-  | {
-      isHW: true
-      cbor: string
-      partial?: boolean
-      onConfirm: (transaction: Transaction) => void
-      onCancel: () => void
-    }
+export type ReviewTransactionParams = {
+  cbor: string
+  onConfirm: () => void
+  onCancel: () => void
+}
 
 export const ReviewTransaction = () => {
   const params = useParams<ReviewTransactionParams>(isParams)
-  const promptRootKey = useConnectorPromptRootKey()
   const [inputsOpen, setInputsOpen] = React.useState(true)
   const [outputsOpen, setOutputsOpen] = React.useState(true)
   const [scrollbarShown, setScrollbarShown] = React.useState(false)
@@ -55,24 +39,6 @@ export const ReviewTransaction = () => {
   const formattedTX = useFormattedTransaction(params.cbor)
 
   const {styles} = useStyles()
-
-  const {sign: signTxWithHW} = useSignTxWithHW()
-
-  const handleOnConfirm = async () => {
-    if (!params.isHW) {
-      const rootKey = await promptRootKey()
-      params.onConfirm(rootKey)
-      return
-    }
-
-    signTxWithHW(
-      {cbor: params.cbor, partial: params.partial},
-      {
-        onSuccess: (signature) => params.onConfirm(signature),
-        onError: (error) => logger.error('ReviewTransaction::handleOnConfirm', {error}),
-      },
-    )
-  }
 
   useEffect(() => {
     return () => {
@@ -138,7 +104,7 @@ export const ReviewTransaction = () => {
       </ScrollView>
 
       <View style={[styles.buttonArea, {borderTopWidth: scrollbarShown ? 1 : 0}]}>
-        <Button title={strings.confirm} onPress={handleOnConfirm} />
+        <Button title={strings.confirm} onPress={params?.onConfirm} />
       </View>
     </SafeAreaView>
   )
@@ -157,21 +123,11 @@ const Dropdown = ({children, open, onPress}: {children: React.ReactNode; open: b
   )
 }
 
-const paramsSchema = z.union([
-  z.object({
-    isHW: z.literal(false),
-    cbor: z.string(),
-    onConfirm: z.function(),
-    onCancel: z.function(),
-  }),
-  z.object({
-    isHW: z.literal(true),
-    cbor: z.string(),
-    partial: z.boolean().optional(),
-    onConfirm: z.function(),
-    onCancel: z.function(),
-  }),
-])
+const paramsSchema = z.object({
+  cbor: z.string(),
+  onConfirm: z.function(),
+  onCancel: z.function(),
+})
 
 const isParams = createTypeGuardFromSchema(paramsSchema)
 
@@ -427,68 +383,4 @@ const FeeChip = () => {
       <Text style={[styles.chipText, {backgroundColor}]}>{strings.fee}</Text>
     </View>
   )
-}
-
-const useConnectorPromptRootKey = () => {
-  const promptRootKey = usePromptRootKey()
-
-  return React.useCallback(() => {
-    return new Promise<string>((resolve, reject) => {
-      let shouldResolveOnClose = true
-
-      try {
-        promptRootKey({
-          onConfirm: (rootKey) => {
-            resolve(rootKey)
-            shouldResolveOnClose = false
-            return Promise.resolve()
-          },
-          onClose: () => {
-            if (shouldResolveOnClose) reject(userRejectedError())
-          },
-        })
-      } catch (error) {
-        reject(error)
-      }
-    })
-  }, [promptRootKey])
-}
-
-export const useSignTxWithHW = () => {
-  const {confirmHWConnection, closeModal} = useConfirmHWConnectionModal()
-  const {wallet, meta} = useSelectedWallet()
-
-  const mutationFn = React.useCallback(
-    (options: {cbor: string; partial?: boolean}) => {
-      return new Promise<Transaction>((resolve, reject) => {
-        let shouldResolveOnClose = true
-        confirmHWConnection({
-          onConfirm: async ({transportType, deviceInfo}) => {
-            try {
-              const cip30 = cip30LedgerExtensionMaker(wallet, meta)
-              const tx = await cip30.signTx(options.cbor, options.partial ?? false, deviceInfo, transportType === 'USB')
-              shouldResolveOnClose = false
-              return resolve(tx)
-            } catch (error) {
-              reject(error)
-            } finally {
-              closeModal()
-            }
-          },
-          onClose: () => {
-            if (shouldResolveOnClose) reject(userRejectedError())
-          },
-        })
-      })
-    },
-    [confirmHWConnection, wallet, meta, closeModal],
-  )
-
-  const mutation = useMutation<Transaction, Error, {cbor: string; partial?: boolean}>({
-    mutationFn,
-    useErrorBoundary: (error) => !isUserRejectedError(error),
-    mutationKey: ['useSignTxWithHW'],
-  })
-
-  return {...mutation, sign: mutation.mutate}
 }
